Changed unsigned ints to size_ts

+++ BitVector.c
@@ -11,9 +11,9 @@
 
 #include "BitVector.h"
 
-void INITIALIZE(BitVector *bv, unsigned int num_bits)
+void INITIALIZE(BitVector *bv, size_t num_bits)
 {
-	unsigned int units = ((num_bits - 1) / BITS_PER_UNIT) + 1;
+	size_t units = ((num_bits - 1) / BITS_PER_UNIT) + 1;
 	bv->bits = calloc(units, sizeof *bv->bits);
 	bv->units = units;
 }
@@ -25,7 +25,7 @@ void FREE_BITVECTOR(BitVector *bv)
 
 void INVERT_ALL_BITS(BitVector *bv)
 {
-	unsigned int i;
+	size_t i;
 	for (i = 0; i < bv->units; i++)
 	{
 		bv->bits[i] ^= MAX_UNIT_VALUE;
@@ -42,34 +42,34 @@ void UNSET_ALL_BITS(BitVector *bv)
 	memset(bv->bits, 0, bv->units * sizeof *bv->bits);
 }
 
-void SET_BIT(BitVector *bv, unsigned int bit_num)
+void SET_BIT(BitVector *bv, size_t bit_num)
 {
-	unsigned int unit = bit_num / BITS_PER_UNIT;
-	unsigned int bit = bit_num % BITS_PER_UNIT;
+	size_t unit = bit_num / BITS_PER_UNIT;
+	size_t bit = bit_num % BITS_PER_UNIT;
 
 	bv->bits[unit] |= 1 << bit;
 }
 
-void UNSET_BIT(BitVector *bv, unsigned int bit_num)
+void UNSET_BIT(BitVector *bv, size_t bit_num)
 {
-	unsigned int unit = bit_num / BITS_PER_UNIT;
-	unsigned int bit = bit_num % BITS_PER_UNIT;
+	size_t unit = bit_num / BITS_PER_UNIT;
+	size_t bit = bit_num % BITS_PER_UNIT;
 
 	bv->bits[unit] &= ~(1 << bit);
 }
 
-void INVERT_BIT(BitVector *bv, unsigned int bit_num)
+void INVERT_BIT(BitVector *bv, size_t bit_num)
 {
-	unsigned int unit = bit_num / BITS_PER_UNIT;
-	unsigned int bit = bit_num % BITS_PER_UNIT;
+	size_t unit = bit_num / BITS_PER_UNIT;
+	size_t bit = bit_num % BITS_PER_UNIT;
 
 	bv->bits[unit] ^= (1 << bit);
 }
 
-unsigned short GET_BIT(BitVector *bv, unsigned int bit_num)
+unsigned short GET_BIT(BitVector *bv, size_t bit_num)
 {
-	unsigned int unit = bit_num / BITS_PER_UNIT;
-	unsigned int bit = bit_num % BITS_PER_UNIT;
+	size_t unit = bit_num / BITS_PER_UNIT;
+	size_t bit = bit_num % BITS_PER_UNIT;
 
 	return bv->bits[unit] & (1 << bit) ? 1 : 0;
 }
